split out remaining next/github api functions into own directory
